+++ src-tauri/src/lib.rs
@@ -392,9 +392,9 @@ pub fn run() {
             audio::commands::audio_resume,
             audio::commands::audio_stop,
             audio::commands::audio_seek,
-            audio::commands::audio_set_volume,
-            audio::commands::audio_update_replay_gain,
-            audio::commands::audio_set_eq,
+            audio::mix_commands::audio_set_volume,
+            audio::mix_commands::audio_update_replay_gain,
+            audio::mix_commands::audio_set_eq,
             audio::autoeq_commands::autoeq_entries,
             audio::autoeq_commands::autoeq_fetch_profile,
             audio::commands::audio_preload,
@@ -403,9 +403,9 @@ pub fn run() {
             audio::preview::audio_preview_stop,
             audio::preview::audio_preview_stop_silent,
             audio::preview::audio_preview_set_volume,
-            audio::commands::audio_set_crossfade,
-            audio::commands::audio_set_gapless,
-            audio::commands::audio_set_normalization,
+            audio::mix_commands::audio_set_crossfade,
+            audio::mix_commands::audio_set_gapless,
+            audio::mix_commands::audio_set_normalization,
             audio::device_commands::audio_list_devices,
             audio::device_commands::audio_canonicalize_selected_device,
             audio::device_commands::audio_default_output_device_name,
